requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
redis==4.6.0
Pillow==10.0.0
email-validator==2.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
Flask-Mail==0.9.1